fix: consolidate get uploader text
