+++ src/components/v5/common/AvatarUploader/hooks.tsx
@@ -11,6 +11,8 @@ import {
 } from '~images/optimisation';
 import { convertBytes } from '~utils/convertBytes';
 
+import { FileUploadOptions } from './types';
+
 export interface UseAvatarUploaderProps {
   updateFn: (
     avatar: string | null,
@@ -93,4 +95,21 @@ export const useFormatFormats = (fileFormats: string[]) => {
       return format;
     })
     .join(', ');
+};
+
+export const useGetUploaderText = (fileOptions: FileUploadOptions) => {
+  const { formatMessage } = useIntl();
+
+  const { fileDimension, fileFormat, fileSize } = fileOptions;
+
+  const formattedFormats = useFormatFormats(fileFormat);
+
+  return formatMessage(
+    { id: 'avatar.uploader.info' },
+    {
+      format: formattedFormats,
+      dimension: fileDimension,
+      size: fileSize,
+    },
+  );
 };
